feat: api routes have network
