+++ __tests__/api/events/cron/processNewOnchainEvents.test.ts
@@ -1,6 +1,6 @@
 import { main } from 'lib/events/sqs/consumer';
 import { createMocks } from 'node-mocks-http';
-import handler from 'pages/api/events/cron/processNewOnchainEvents';
+import handler from 'pages/api/network/[network]/events/cron/processNewOnchainEvents';
 
 jest.mock('lib/events/sqs/consumer', () => ({
   main: jest.fn(),
