+++ __tests__/api/events/updateDiscordMetrics.test.ts
@@ -7,7 +7,7 @@ import {
 } from 'lib/loans/subgraph/subgraphLoans';
 import { subgraphLoan } from 'lib/mockData';
 import { createMocks } from 'node-mocks-http';
-import handler from 'pages/api/events/updateDiscordMetrics';
+import handler from 'pages/api/network/[network]/events/updateDiscordMetrics';
 
 jest.mock('lib/events/consumers/discord/bot', () => ({
   updateWatcher: jest.fn(),
