+++ pages/api/network/[network]/addresses/[address]/notifications/emails/[emailOrUuid].ts
