+++ pages/api/network/[network]/events/_middleware.ts
