+++ pages/api/network/[network]/loanAssets.ts
