+++ pages/api/network/[network]/loans/[id].tsx
