+++ pages/api/network/[network]/loans/search.ts
